pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "webstatuspi"
version = "0.1.0"
description = "Ultra-lightweight web monitoring system for Raspberry Pi 1B+"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Jose Manuel Lucas Munoz"}
]
keywords = ["monitoring", "raspberry-pi", "web-status", "status-tracker"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
    "coverage>=6.0.0",
]

[project.scripts]
webstatuspi = "webstatuspi:main"

[tool.setuptools.packages.find]
where = ["."]
